+++ packages/core/admin/admin/src/content-manager/components/ConfigurationForm/Fields.tsx
@@ -9,8 +9,8 @@ import { useIntl } from 'react-intl';
 import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
 
+import { ItemTypes } from '../../constants/dragAndDrop';
 import { type UseDragAndDropOptions, useDragAndDrop } from '../../hooks/useDragAndDrop';
-import { ItemTypes } from '../../utils/dragAndDrop';
 import { useComposedRefs } from '../../utils/refs';
 import { getTranslation } from '../../utils/translations';
 import { ComponentIcon } from '../ComponentIcon';
