+++ packages/core/admin/admin/src/content-manager/components/Relations/RelationInput.tsx
@@ -23,13 +23,13 @@ import { useIntl } from 'react-intl';
 import { FixedSizeList, FixedSizeList as List, ListChildComponentProps } from 'react-window';
 import styled from 'styled-components';
 
+import { ItemTypes } from '../../constants/dragAndDrop';
 import {
   UseDragAndDropOptions,
   useDragAndDrop,
   DROP_SENSITIVITY,
 } from '../../hooks/useDragAndDrop';
 import { usePrev } from '../../hooks/usePrev';
-import { ItemTypes } from '../../utils/dragAndDrop';
 import { composeRefs } from '../../utils/refs';
 import { getTranslation } from '../../utils/translations';
