+++ packages/core/admin/admin/src/content-manager/constants/dragAndDrop.ts
