+++ packages/core/admin/admin/src/content-manager/pages/EditView/components/FormInputs/BlocksInput/BlocksContent.tsx
@@ -7,8 +7,8 @@ import { Editor, Range, Transforms } from 'slate';
 import { ReactEditor, type RenderElementProps, type RenderLeafProps, Editable } from 'slate-react';
 import styled, { CSSProperties, css } from 'styled-components';
 
+import { ItemTypes } from '../../../../../constants/dragAndDrop';
 import { useDragAndDrop, DIRECTIONS } from '../../../../../hooks/useDragAndDrop';
-import { ItemTypes } from '../../../../../utils/dragAndDrop';
 import { composeRefs } from '../../../../../utils/refs';
 import { getTranslation } from '../../../../../utils/translations';
